3.731 (05/01/2010)
Document TeamBalancer and Steam lobby plugins. Fix crash with calling RakPeer::DetachPlugin(). Fix bug to return ID_SND_RECEIPT_ACKED only once for messages over the MTU. Rewrite RakNetStatistics to return more relevant and up to date information, including metrics per second. Add VariableDeltaSerializer, an optional helper class with ReplicaManager3 that can serialize changes per-variable, rather than relying on the default of serializing the entire object.

+++ Samples/Tests/ConnectWithSocketTest.cpp
@@ -1,7 +1,5 @@
 #include "ConnectWithSocketTest.h"
 
-
-
 /*
 Description:
 virtual bool RakPeerInterface::ConnectWithSocket  	(  	const char *   	 host, 		unsigned short  	remotePort, 		const char *  	passwordData, 		int  	passwordDataLength, 		RakNetSmartPtr< RakNetSocket >  	socket, 		unsigned  	sendConnectionAttemptCount = 7, 		unsigned  	timeBetweenSendConnectionAttemptsMS = 500, 		RakNetTime  	timeoutTime = 0	  	) 	
@@ -11,12 +9,7 @@ virtual RakNetSmartPtr<RakNetSocket> RakPeerInterface::GetSocket  	(  	const Sys
 
 Success conditions:
 
-
-
-
 Failure conditions:
-
-
 
 RakPeerInterface Functions used, tested indirectly by its use:
 Startup
@@ -36,27 +29,20 @@ int ConnectWithSocketTest::RunTest(DataStructures::List<RakNet::RakString> param
 {
 	destroyList.Clear(false,__FILE__,__LINE__);
 
-
 	RakPeerInterface *server,*client;
 
 	DataStructures::List< RakNetSmartPtr< RakNetSocket > > sockets;
 	TestHelpers::StandardClientPrep(client,destroyList);
 	TestHelpers::StandardServerPrep(server,destroyList);
 
-
-
 	SystemAddress serverAddress;
 
 	serverAddress.SetBinaryAddress("127.0.0.1");
 	serverAddress.port=60000;
 
-
-
-
 	printf("Testing normal connect before test\n");
 	if (!TestHelpers::WaitAndConnectTwoPeersLocally(client,server,5000))
 	{
-
 
 		if (isVerbose)
 			DebugTools::ShowError(errorList[1-1],!noPauses && isVerbose,__LINE__,__FILE__);
@@ -68,7 +54,6 @@ int ConnectWithSocketTest::RunTest(DataStructures::List<RakNet::RakString> param
 
 	if (!TestHelpers::WaitForTestPacket(server,5000))
 	{
-
 
 		if (isVerbose)
 			DebugTools::ShowError(errorList[2-1],!noPauses && isVerbose,__LINE__,__FILE__);
@@ -83,18 +68,13 @@ int ConnectWithSocketTest::RunTest(DataStructures::List<RakNet::RakString> param
 
 	client->GetSockets(sockets);
 
-
 	theSocket=sockets[0];
-
-
-
 
 	RakTimer timer2(5000);
 
 	printf("Testing ConnectWithSocket using socket from GetSockets\n");
 	while(!client->IsConnected (serverAddress,false,false)&&!timer2.IsExpired())
 	{
-
 
 		if(!client->IsConnected (serverAddress,true,true))
 		{
@@ -105,12 +85,8 @@ int ConnectWithSocketTest::RunTest(DataStructures::List<RakNet::RakString> param
 
 	}
 
-
-
 	if (!client->IsConnected (serverAddress,false,false))
 	{
-
-
 
 		if (isVerbose)
 			DebugTools::ShowError(errorList[3-1],!noPauses && isVerbose,__LINE__,__FILE__);
@@ -120,11 +96,8 @@ int ConnectWithSocketTest::RunTest(DataStructures::List<RakNet::RakString> param
 
 	TestHelpers::BroadCastTestPacket(client);
 
-
 	if (!TestHelpers::WaitForTestPacket(server,5000))
 	{
-
-
 
 		if (isVerbose)
 			DebugTools::ShowError(errorList[4-1],!noPauses && isVerbose,__LINE__,__FILE__);
@@ -133,12 +106,8 @@ int ConnectWithSocketTest::RunTest(DataStructures::List<RakNet::RakString> param
 
 	}
 
-
-
-
 	printf("Disconnecting client\n");
 	CommonFunctions::DisconnectAndWait(client,"127.0.0.1",60000);
-
 
 	printf("Testing ConnectWithSocket using socket from GetSocket\n");
 	theSocket=client->GetSocket(UNASSIGNED_SYSTEM_ADDRESS);//Get open Socket
@@ -147,7 +116,6 @@ int ConnectWithSocketTest::RunTest(DataStructures::List<RakNet::RakString> param
 
 	while(!client->IsConnected (serverAddress,false,false)&&!timer2.IsExpired())
 	{
-
 
 		if(!client->IsConnected (serverAddress,true,true))
 		{
@@ -158,12 +126,8 @@ int ConnectWithSocketTest::RunTest(DataStructures::List<RakNet::RakString> param
 
 	}
 
-
-
 	if (!client->IsConnected (serverAddress,false,false))
 	{
-
-
 
 		if (isVerbose)
 			DebugTools::ShowError(errorList[5-1],!noPauses && isVerbose,__LINE__,__FILE__);
@@ -171,15 +135,10 @@ int ConnectWithSocketTest::RunTest(DataStructures::List<RakNet::RakString> param
 		return 5;
 	}
 
-
-
 	TestHelpers::BroadCastTestPacket(client);
-
 
 	if (!TestHelpers::WaitForTestPacket(server,5000))
 	{
-
-
 
 		if (isVerbose)
 			DebugTools::ShowError(errorList[6-1],!noPauses && isVerbose,__LINE__,__FILE__);
@@ -192,10 +151,6 @@ int ConnectWithSocketTest::RunTest(DataStructures::List<RakNet::RakString> param
 
 }
 
-
-
-
-
 RakNet::RakString ConnectWithSocketTest::GetTestName()
 {
 
@@ -206,7 +161,7 @@ RakNet::RakString ConnectWithSocketTest::GetTestName()
 RakNet::RakString ConnectWithSocketTest::ErrorCodeToString(int errorCode)
 {
 
-	if (errorCode>0&&errorCode<=errorList.Size())
+	if (errorCode>0&&(unsigned int)errorCode<=errorList.Size())
 	{
 		return errorList[errorCode-1];
 	}
@@ -215,10 +170,7 @@ RakNet::RakString ConnectWithSocketTest::ErrorCodeToString(int errorCode)
 		return "Undefined Error";
 	}	
 
-
 }
-
-
 
 ConnectWithSocketTest::ConnectWithSocketTest(void)
 {
@@ -234,7 +186,6 @@ ConnectWithSocketTest::ConnectWithSocketTest(void)
 ConnectWithSocketTest::~ConnectWithSocketTest(void)
 {
 }
-
 
 void ConnectWithSocketTest::DestroyPeers()
 {
