+++ Samples/Tests/MiscellaneousTestsTest.cpp
@@ -1,7 +1,5 @@
 #include "MiscellaneousTestsTest.h"
 
-
-
 /*
 Description:
 Tests:
@@ -9,16 +7,9 @@ virtual void 	SetRouterInterface (RouterInterface *routerInterface)=0
 virtual void 	RemoveRouterInterface (RouterInterface *routerInterface)=0
 virtual bool 	AdvertiseSystem (const char *host, unsigned short remotePort, const char *data, int dataLength, unsigned connectionSocketIndex=0)=0
 
-
-
 Success conditions:
 
-
-
-
 Failure conditions:
-
-
 
 RakPeerInterface Functions used, tested indirectly by its use,list may not be complete:
 Startup
@@ -32,19 +23,14 @@ SetRouterInterface
 RemoveRouterInterface
 AdvertiseSystem
 
-
 */
 int MiscellaneousTestsTest::RunTest(DataStructures::List<RakNet::RakString> params,bool isVerbose,bool noPauses)
 {	destroyList.Clear(false,__FILE__,__LINE__);
-
 
 RakPeerInterface *client,*server;
 
 TestHelpers::StandardClientPrep(client,destroyList);
 TestHelpers::StandardServerPrep(server,destroyList);
-
-
-
 
 printf("Testing AdvertiseSystem\n");
 
@@ -53,16 +39,11 @@ client->AdvertiseSystem("127.0.0.1",60000,0,0);
 if (!CommonFunctions::WaitForMessageWithID(server,ID_ADVERTISE_SYSTEM,5000))
 {
 
-
-
 	if (isVerbose)
 		DebugTools::ShowError(errorList[1-1],!noPauses && isVerbose,__LINE__,__FILE__);
 
 	return 1;
 }
-
-
-
 
 RouterInterfaceTester * riTester= new RouterInterfaceTester();
 
@@ -70,12 +51,8 @@ printf("Testing SetRouterInterface\n");
 
 client->SetRouterInterface(riTester);
 
-
-
 if (riTester->wasTriggered())
 {
-
-
 
 	if (isVerbose)
 		DebugTools::ShowError(errorList[2-1],!noPauses && isVerbose,__LINE__,__FILE__);
@@ -83,13 +60,10 @@ if (riTester->wasTriggered())
 	return 2;
 }
 
-
 TestHelpers::SendTestPacketDirected(client,"127.0.0.1",60001);
 
 if (!riTester->wasTriggered())
 {
-
-
 
 	if (isVerbose)
 		DebugTools::ShowError(errorList[3-1],!noPauses && isVerbose,__LINE__,__FILE__);
@@ -106,26 +80,15 @@ TestHelpers::SendTestPacketDirected(client,"127.0.0.1",60001);
 if (riTester->wasTriggered())
 {
 
-
-
 	if (isVerbose)
 		DebugTools::ShowError(errorList[4-1],!noPauses && isVerbose,__LINE__,__FILE__);
 
 	return 4;
 }
 
-
-
-
-
-
-
-
 return 0;
 
 }
-
-
 
 RakNet::RakString MiscellaneousTestsTest::GetTestName()
 {
@@ -137,7 +100,7 @@ RakNet::RakString MiscellaneousTestsTest::GetTestName()
 RakNet::RakString MiscellaneousTestsTest::ErrorCodeToString(int errorCode)
 {
 
-	if (errorCode>0&&errorCode<=errorList.Size())
+	if (errorCode>0&&(unsigned int)errorCode<=errorList.Size())
 	{
 		return errorList[errorCode-1];
 	}
@@ -146,9 +109,7 @@ RakNet::RakString MiscellaneousTestsTest::ErrorCodeToString(int errorCode)
 		return "Undefined Error";
 	}	
 
-
 }
-
 
 void MiscellaneousTestsTest::DestroyPeers()
 {
@@ -160,8 +121,6 @@ void MiscellaneousTestsTest::DestroyPeers()
 
 }
 
-
-
 MiscellaneousTestsTest::MiscellaneousTestsTest(void)
 {
 
@@ -169,7 +128,6 @@ MiscellaneousTestsTest::MiscellaneousTestsTest(void)
 	errorList.Push("The router interface should not be called because no send has happened yet",__FILE__,__LINE__);
 	errorList.Push("Router failed to trigger on failed directed send",__FILE__,__LINE__);
 	errorList.Push("Router was not properly removed",__FILE__,__LINE__);
-
 
 }
