+++ Samples/Tests/PacketAndLowLevelTestsTest.cpp
@@ -2,7 +2,6 @@
 
 #include "malloc.h"
 
-
 /*
 Description:
 Tests out the sunctions:
@@ -17,12 +16,7 @@ DetachPlugin (PluginInterface2 *plugin)=0
 
 Success conditions:
 
-
-
-
 Failure conditions:
-
-
 
 RakPeerInterface Functions used, tested indirectly by its use,list may not be complete:
 Startup
@@ -53,14 +47,11 @@ int PacketAndLowLevelTestsTest::RunTest(DataStructures::List<RakNet::RakString> 
 	if (!TestHelpers::WaitAndConnectTwoPeersLocally(client,server,5000))
 	{
 
-
-
 		if (isVerbose)
 			DebugTools::ShowError(errorList[1-1],!noPauses && isVerbose,__LINE__,__FILE__);
 
 		return 1;
 	}
-
 
 	printf("Testing SendList\n");
 
@@ -73,14 +64,11 @@ int PacketAndLowLevelTestsTest::RunTest(DataStructures::List<RakNet::RakString> 
 	char curString4[]="ADDD";
 	char curString5[]="AEEE";
 
-
 	dataList[0]=curString1;
 	dataList[1]=curString2;
 	dataList[2]=curString3;
 	dataList[3]=curString4;
 	dataList[4]=curString5;
-
-
 
 	for (int i=0;i<5;i++)
 	{
@@ -90,20 +78,15 @@ int PacketAndLowLevelTestsTest::RunTest(DataStructures::List<RakNet::RakString> 
 
 	client->SendList((const char**)dataList,lengths,5,HIGH_PRIORITY,RELIABLE_ORDERED,0, UNASSIGNED_SYSTEM_ADDRESS, true);
 
-
-
 	Packet* packet;
 	if (!(packet=CommonFunctions::WaitAndReturnMessageWithID(server,ID_USER_PACKET_ENUM+1,1000)))
 	{
-
-
 
 		if (isVerbose)
 			DebugTools::ShowError(errorList[9-1],!noPauses && isVerbose,__LINE__,__FILE__);
 
 		return 9;
 	}
-
 
 	if (packet->length!=25)
 	{
@@ -119,13 +102,11 @@ int PacketAndLowLevelTestsTest::RunTest(DataStructures::List<RakNet::RakString> 
 
 	PluginInterface2* myPlug=new PacketChangerPlugin();
 
-
 	printf("Test attach detach of plugins\n");
 	client->AttachPlugin(myPlug); 
 	TestHelpers::BroadCastTestPacket(client);
 	if (TestHelpers::WaitForTestPacket(server,2000))
 	{
-
 
 		if (isVerbose)
 			DebugTools::ShowError(errorList[2-1],!noPauses && isVerbose,__LINE__,__FILE__);
@@ -137,7 +118,6 @@ int PacketAndLowLevelTestsTest::RunTest(DataStructures::List<RakNet::RakString> 
 	TestHelpers::BroadCastTestPacket(client);
 	if (!TestHelpers::WaitForTestPacket(server,2000))
 	{
-
 
 		if (isVerbose)
 			DebugTools::ShowError(errorList[3-1],!noPauses && isVerbose,__LINE__,__FILE__);
@@ -151,7 +131,6 @@ int PacketAndLowLevelTestsTest::RunTest(DataStructures::List<RakNet::RakString> 
 	hugePacket=client->AllocatePacket(dataSize);
 	hugePacket2=client->AllocatePacket(dataSize);
 
-
 	/*//Couldn't find a good cross platform way for allocated memory so skipped this check
 	if (somemalloccheck<3000000)
 	{}
@@ -159,18 +138,13 @@ int PacketAndLowLevelTestsTest::RunTest(DataStructures::List<RakNet::RakString> 
 
 	printf("Assuming 3000000 allocation for splitpacket, testing setsplitpacket\n");
 
-
 	hugePacket->data[0]=ID_USER_PACKET_ENUM+1;
 	hugePacket2->data[0]=ID_USER_PACKET_ENUM+1;
 
-
 	server->SetSplitMessageProgressInterval(1);
-
 
 	if (server->GetSplitMessageProgressInterval()!=1)
 	{
-
-
 
 		if (isVerbose)
 			DebugTools::ShowError(errorList[4-1],!noPauses && isVerbose,__LINE__,__FILE__);
@@ -179,11 +153,8 @@ int PacketAndLowLevelTestsTest::RunTest(DataStructures::List<RakNet::RakString> 
 
 	}
 
-
 	if (!client->Send((const char *)hugePacket->data,dataSize,HIGH_PRIORITY,RELIABLE_ORDERED,0, UNASSIGNED_SYSTEM_ADDRESS, true))
 	{
-
-
 
 		if (isVerbose)
 			DebugTools::ShowError(errorList[5-1],!noPauses && isVerbose,__LINE__,__FILE__);
@@ -194,26 +165,19 @@ int PacketAndLowLevelTestsTest::RunTest(DataStructures::List<RakNet::RakString> 
 	if (!CommonFunctions::WaitForMessageWithID(server,ID_DOWNLOAD_PROGRESS,2000))
 	{
 
-
-
 		if (isVerbose)
 			DebugTools::ShowError(errorList[6-1],!noPauses && isVerbose,__LINE__,__FILE__);
 
 		return 6;
 	}
 
-
-
 	while(CommonFunctions::WaitForMessageWithID(server,ID_DOWNLOAD_PROGRESS,500))//Clear out the rest before next test
 	{
 	}
 
-
 	printf("Making sure still connected, if not connect\n");
 	if (!TestHelpers::WaitAndConnectTwoPeersLocally(client,server,5000))//Make sure connected before test
 	{
-
-
 
 		if (isVerbose)
 			DebugTools::ShowError(errorList[11-1],!noPauses && isVerbose,__LINE__,__FILE__);
@@ -221,12 +185,10 @@ int PacketAndLowLevelTestsTest::RunTest(DataStructures::List<RakNet::RakString> 
 		return 11;
 	}
 
-
 	printf("Making sure standard send/recieve still functioning\n");
 	TestHelpers::BroadCastTestPacket(client);
 	if (!TestHelpers::WaitForTestPacket(server,5000))
 	{
-
 
 		if (isVerbose)
 			DebugTools::ShowError(errorList[12],!noPauses && isVerbose,__LINE__,__FILE__);
@@ -238,11 +200,8 @@ int PacketAndLowLevelTestsTest::RunTest(DataStructures::List<RakNet::RakString> 
 
 	server->PushBackPacket(hugePacket,false);
 
-
-
 	if (!TestHelpers::WaitForTestPacket(server,2000))
 	{
-
 
 		if (isVerbose)
 			DebugTools::ShowError(errorList[7-1],!noPauses && isVerbose,__LINE__,__FILE__);
@@ -250,13 +209,9 @@ int PacketAndLowLevelTestsTest::RunTest(DataStructures::List<RakNet::RakString> 
 		return 7;
 	}
 
-
-
 	printf("Making sure still connected, if not connect\n");
 	if (!TestHelpers::WaitAndConnectTwoPeersLocally(client,server,5000))//Make sure connected before test
 	{
-
-
 
 		if (isVerbose)
 			DebugTools::ShowError(errorList[11-1],!noPauses && isVerbose,__LINE__,__FILE__);
@@ -264,12 +219,10 @@ int PacketAndLowLevelTestsTest::RunTest(DataStructures::List<RakNet::RakString> 
 		return 11;
 	}
 
-
 	printf("Making sure standard send/recieve still functioning\n");
 	TestHelpers::BroadCastTestPacket(client);
 	if (!TestHelpers::WaitForTestPacket(server,2000))
 	{
-
 
 		if (isVerbose)
 			DebugTools::ShowError(errorList[12-1],!noPauses && isVerbose,__LINE__,__FILE__);
@@ -277,19 +230,11 @@ int PacketAndLowLevelTestsTest::RunTest(DataStructures::List<RakNet::RakString> 
 		return 12;
 	}
 
-
-
-
-
-
 	printf("PushBackPacket head true test\n");
 	server->PushBackPacket(hugePacket2,true);
 
-
-
 	if (!TestHelpers::WaitForTestPacket(server,2000))
 	{
-
 
 		if (isVerbose)
 			DebugTools::ShowError(errorList[10-1],!noPauses && isVerbose,__LINE__,__FILE__);
@@ -297,12 +242,9 @@ int PacketAndLowLevelTestsTest::RunTest(DataStructures::List<RakNet::RakString> 
 		return 10;
 	}
 
-
 	printf("Making sure still connected, if not connect\n");
 	if (!TestHelpers::WaitAndConnectTwoPeersLocally(client,server,5000))//Make sure connected before test
 	{
-
-
 
 		if (isVerbose)
 			DebugTools::ShowError(errorList[11-1],!noPauses && isVerbose,__LINE__,__FILE__);
@@ -310,20 +252,16 @@ int PacketAndLowLevelTestsTest::RunTest(DataStructures::List<RakNet::RakString> 
 		return 11;
 	}
 
-
 	printf("Run recieve test\n");
 	TestHelpers::BroadCastTestPacket(client);
 	if (!TestHelpers::WaitForTestPacket(server,2000))
 	{
-
 
 		if (isVerbose)
 			DebugTools::ShowError(errorList[12-1],!noPauses && isVerbose,__LINE__,__FILE__);
 
 		return 12;
 	}
-
-
 
 	return 0;
 
@@ -339,7 +277,6 @@ void PacketAndLowLevelTestsTest::FloodWithHighPriority(RakPeerInterface* client)
 
 }
 
-
 RakNet::RakString PacketAndLowLevelTestsTest::GetTestName()
 {
 
@@ -350,7 +287,7 @@ RakNet::RakString PacketAndLowLevelTestsTest::GetTestName()
 RakNet::RakString PacketAndLowLevelTestsTest::ErrorCodeToString(int errorCode)
 {
 
-	if (errorCode>0&&errorCode<=errorList.Size())
+	if (errorCode>0&&(unsigned int)errorCode<=errorList.Size())
 	{
 		return errorList[errorCode-1];
 	}
@@ -359,9 +296,7 @@ RakNet::RakString PacketAndLowLevelTestsTest::ErrorCodeToString(int errorCode)
 		return "Undefined Error";
 	}	
 
-
 }
-
 
 void PacketAndLowLevelTestsTest::DestroyPeers()
 {
@@ -372,8 +307,6 @@ void PacketAndLowLevelTestsTest::DestroyPeers()
 		RakNetworkFactory::DestroyRakPeerInterface(destroyList[i]);
 
 }
-
-
 
 PacketAndLowLevelTestsTest::PacketAndLowLevelTestsTest(void)
 {
